add a lot of function

+++ dictionary/functions/suppliers_upd.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION dictionary.suppliers_upd(_src JSONB, _ch_employee_id INT) RETURNS JSONB
+CREATE OR REPLACE FUNCTION dictionary.suppliers_upd(_src JSONB, _ch_employee_id BIGINT) RETURNS JSONB
     SECURITY DEFINER
     LANGUAGE plpgsql
 AS
@@ -51,4 +51,3 @@ BEGIN
     RETURN JSONB_BUILD_OBJECT('data', NULL);
 END
 $$;
-
